+++ DataAccess/Models/Product/ProductModel.cs
@@ -5,11 +5,11 @@ namespace ECommerce.DataAccess.Models.Product
         public int ProductId { get; set; }
         public string Name { get; set; }
         public string Description { get; set; }
-        public double Price { get; set; }
+        public decimal Price { get; set; }
         public int ManufacturerId { get; set; }
         public int CategoryId { get; set; }
 
-        public ProductModel(string name, string description, double price, int manufacturerId, int categoryId)
+        public ProductModel(string name, string description, decimal price, int manufacturerId, int categoryId)
         {
             ProductId = -1;
             Name = name;
@@ -19,7 +19,7 @@ namespace ECommerce.DataAccess.Models.Product
             CategoryId = categoryId;
         }
 
-        public ProductModel(int productId, string name, string description, double price, int manufacturerId, int categoryId)
+        public ProductModel(int productId, string name, string description, decimal price, int manufacturerId, int categoryId)
         {
             ProductId = productId;
             Name = name;
