+++ DataAccess/Repositories/ProductRepository.cs
@@ -102,7 +102,7 @@ namespace ECommerce.DataAccess.Repositories
                     row.GetColumn<int>("productid"),
                     row.GetColumn<string>("name"),
                     row.GetColumn<string>("description"),
-                    (double)row.GetColumn<decimal>("price"), // TODO: decimal
+                    row.GetColumn<decimal>("price"), // TODO: decimal
                     row.GetColumn<int>("manufacturerid"),
                     row.GetColumn<int>("categoryid")
                 );
@@ -126,7 +126,7 @@ namespace ECommerce.DataAccess.Repositories
                     row.GetColumn<int>("productid"),
                     row.GetColumn<string>("name"),
                     row.GetColumn<string>("description"),
-                    (double)row.GetColumn<decimal>("price"), // TODO: decimal
+                    row.GetColumn<decimal>("price"), // TODO: decimal
                     row.GetColumn<int>("manufacturerid"),
                     row.GetColumn<int>("categoryid")
                 ));
